-reverse all client hud panel vtables -increase max edict limit -fix crash to desktop with CL_ParseTEnt rebuild -fix escape menu not functioning correctly -show mission scores when pressing F1 key (temp) -fix mp_timelimit variable type handling and other cvars in amxmodx -crashfix in LoadAdjacentEntities -more function labelling

+++ amx/amxmodx/CTask.cpp
@@ -193,7 +193,7 @@ CTaskMngr::~CTaskMngr()
 	clear();
 }
 
-void CTaskMngr::registerTimers(float *pCurrentTime, float *pTimeLimit, float *pTimeLeft)
+void CTaskMngr::registerTimers(float *pCurrentTime, int *pTimeLimit, float *pTimeLeft)
 {
 	m_pTmr_CurrentTime = pCurrentTime;
 	m_pTmr_TimeLimit = pTimeLimit;
